Changed package names to match project renaming. Updated Kotlin & gradle plugins.
