+++ README.md
@@ -1,15 +1,15 @@
-[logo]: https://www.xy.company/img/home/logo_xy.png "XYAccount"
+[logo]: https://cdn.xy.company/img/brand/XY_Logo_GitHub.png
 
-![logo]
+# ![logo]
 
-# sdk-core-android
+# sdk-base-android
 
-[![](https://jitpack.io/v/XYOracleNetwork/sdk-core-android.svg)](https://jitpack.io/#XYOracleNetwork/sdk-core-android) [![Maintainability](https://api.codeclimate.com/v1/badges/6f2827f8a4103feb7aca/maintainability)](https://codeclimate.com/github/XYOracleNetwork/sdk-core-android/maintainability)
+[![](https://jitpack.io/v/XYOracleNetwork/sdk-base-android.svg)](https://jitpack.io/#XYOracleNetwork/sdk-base-android) [![Maintainability](https://api.codeclimate.com/v1/badges/6f2827f8a4103feb7aca/maintainability)](https://codeclimate.com/github/XYOracleNetwork/sdk-base-android/maintainability)
 
 | Branches        | Status           |
 | ------------- |:-------------:|
-| Master      | [![Build Status](https://travis-ci.com/XYOracleNetwork/sdk-core-android.svg?branch=master)](https://travis-ci.com/XYOracleNetwork/sdk-core-android) |
-| Develop      | [![Build Status](https://travis-ci.com/XYOracleNetwork/sdk-core-android.svg?branch=develop)](https://travis-ci.com/XYOracleNetwork/sdk-core-android)      |
+| Master      | [![Build Status](https://travis-ci.com/XYOracleNetwork/sdk-base-android.svg?branch=master)](https://travis-ci.com/XYOracleNetwork/sdk-base-android) |
+| Develop      | [![Build Status](https://travis-ci.com/XYOracleNetwork/sdk-base-android.svg?branch=develop)](https://travis-ci.com/XYOracleNetwork/sdk-base-android)      |
 
 An XY inspired utility framework for Android and Kotlin
 
@@ -28,7 +28,7 @@ allprojects {
 
 ```gradle
 dependencies {
-    implementation 'com.github.XYOracleNetwork:sdk-core-android:1.0.2'
+    implementation 'com.github.XYOracleNetwork:sdk-base-android:1.2.4'
 }
 ```
 
@@ -53,11 +53,13 @@ Get the device name:
 Logging:
 
 ``` kotlin
- logInfo("A simple INFO log")
- logExtreme("A simple Verbose log")
- logError("An error occured" , false) //set true for debugging
- logException(myException, false) // set true for debugging
- logAction("Log an event or action generated by a user")
+ log.info("A simple INFO log")
+ log.info("myFunctionName", "A simple INFO log")
+ log.action("Actions are events that are generated by the user, like pushing a button")
+ log.status("Status are Large Scale Events, Such As Startup, or Shutdown")
+ log.error("An error occured" , false) //set true to rethrow error
+ log.error(myException) //Any Handled Exceptions
+ log.error("An error with a custom message")
 ```
 
 ## License
