+++ base-android-library/.classpath
