+++ base-android-library/.gitignore
