+++ base-android-sample/.classpath
