+++ base-android-sample/proguard-rules.pro
