+++ base-android-sample/src/main/res/mipmap-xxhdpi/ic_launcher_round.png
