+++ base-android-sample/src/main/res/mipmap-xxxhdpi/ic_launcher.png
