+++ build.gradle
@@ -11,7 +11,7 @@ buildscript {
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:3.3.2'
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.3.21"
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.3.31"
         classpath "com.github.dcendents:android-maven-gradle-plugin:2.1"
 
         // NOTE: Do not place your application dependencies here; they belong
