--- core-android-library/.settings/org.eclipse.buildship.core.prefs
@@ -1,2 +1,0 @@
-connection.project.dir=../../app-xyfindit-android
-eclipse.preferences.version=1
