--- core-android-library/build.gradle
@@ -1,78 +1,0 @@
-apply plugin: 'com.android.library'
-apply plugin: 'kotlin-android'
-apply plugin: 'com.github.dcendents.android-maven'
-group = 'com.github.XYOracleNetwork'
-
-buildscript {
-    repositories {
-        mavenCentral()
-        mavenLocal()
-        jcenter()
-        google()
-    }
-
-    dependencies {
-        classpath 'com.android.tools.build:gradle:3.3.2'
-        classpath 'com.github.dcendents:android-maven-gradle-plugin:2.1'
-    }
-}
-
-android {
-    compileSdkVersion 28
-    buildToolsVersion '28.0.3'
-
-    defaultConfig {
-        minSdkVersion 21
-        targetSdkVersion 28
-        defaultPublishConfig 'release'
-    }
-
-    lintOptions {
-        htmlReport true
-        abortOnError true
-        warningsAsErrors true
-    }
-
-    buildTypes {
-        release {
-            debuggable false
-            minifyEnabled true
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-        debug {
-            minifyEnabled false
-            debuggable true
-        }
-    }
-    compileOptions {
-        kotlinOptions {
-            allWarningsAsErrors = true
-        }
-    }
-
-    sourceSets {
-        main.java.srcDirs += 'src/main/kotlin'
-    }
-
-    packagingOptions {
-        exclude 'META-INF/build.kotlin_module'
-        exclude 'META-INF/atomicfu.kotlin_module'
-        exclude 'META-INF/library_release.kotlin_module'
-    }
-}
-
-repositories {
-    mavenCentral()
-    mavenLocal()
-    jcenter()
-    google()
-}
-
-dependencies {
-    implementation fileTree(include: ['*.jar'], dir: 'libs')
-    api "org.jetbrains.kotlin:kotlin-stdlib-jdk8:1.3.30"
-    api "org.jetbrains.kotlin:kotlin-reflect:1.3.30"
-    api 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.2.0'
-    api 'androidx.appcompat:appcompat:1.1.0-alpha04'
-    api "com.jaredrummler:android-device-names:1.1.8"
-}
