--- core-android-library/src/main/AndroidManifest.xml
@@ -1,8 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="network.xyo.core"
-    android:versionCode='10203'
-    android:versionName="1.2.3">
-
-    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
-
-</manifest>
