--- core-android-library/src/main/kotlin/network/xyo/core/XYLogging.kt
@@ -1,70 +1,0 @@
-package network.xyo.core
-
-import android.util.Log
-
-/* We have everyone of these functions returning 'this' to allow for chaining */
-
-open class XYLogging(private val source: String) {
-    fun error(message: String, reThrow: Boolean): XYLogging {
-        Log.e(info.sourceNameFromAny(source), message)
-        Log.e(info.sourceNameFromAny(source), Thread.currentThread().stackTrace?.contentToString()?.replace(", ", ",\r\n")
-                ?: "Unknown Thread")
-        if (reThrow) {
-            throw RuntimeException()
-        }
-        return this
-    }
-
-    //Any Handled Exceptions
-    fun error(ex: Throwable): XYLogging {
-        error(ex, false)
-        return this
-    }
-
-    fun error(ex: Throwable, reThrow: Boolean): XYLogging {
-        Log.e(info.sourceNameFromAny(source), info.classNameFromObject(ex))
-        ex.stackTrace?.let {stackTrace ->
-            android.util.Log.e(info.sourceNameFromAny(source), stackTrace.contentToString().replace(", ", ",\r\n"))
-        }
-
-        if (info.hasDebugger) {
-            if (reThrow) {
-                throw RuntimeException(ex)
-            }
-        }
-        return this
-    }
-
-    fun error(message: String): XYLogging {
-        Log.e(info.sourceNameFromAny(source), "$message:${info.currentThreadName}")
-        return this
-    }
-
-    //Normal information used for debugging.  Items should be less noisy than Extreme items
-    fun info(function: String, message: String): XYLogging {
-        Log.i(info.sourceNameFromAny(source), "$source:$function:$message [${info.currentThreadName}]")
-        return this
-    }
-
-    fun info(message: String): XYLogging {
-        Log.i(info.sourceNameFromAny(source), "$message [${info.currentThreadName}]")
-        return this
-    }
-
-    //Actions are events that are generated by the user, like pushing a button
-    fun action(action: String): XYLogging {
-        Log.i(info.sourceNameFromAny(source), action)
-        return this
-    }
-
-    //Status are Large Scale Events, Such As Startup, or Shutdown,
-    //that may or may not be a result of a user action
-    fun status(status: String): XYLogging {
-        Log.i(info.sourceNameFromAny(source), "App Status: $status")
-        return this
-    }
-
-    companion object {
-        val info = XYInfo()
-    }
-}
